+++ src/relife/lifetime_models/_distributions.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from collections.abc import Callable
+from abc import ABC
+from collections.abc import Callable, Sequence
 from typing import (
     Any,
     Concatenate,
@@ -15,22 +15,26 @@ from typing import (
 import numpy as np
 import numpydoc.docscrape as docscrape  # pyright: ignore[reportMissingTypeStubs]
 from optype.numpy import Array, Array1D, ArrayND
+from scipy.optimize import Bounds
 from scipy.special import digamma, exp1, gamma, gammaincc, gammainccinv
 from typing_extensions import override
 
-from relife.base import FittingResults
+from relife.base import FitConfig, FittingResults
 from relife.quadratures import (
     laguerre_quadrature,
 )
 from relife.typing import VT
 
 from ._base import (
+    FittableParametricLifetimeModel,
+    LifetimeData,
+    LifetimeLikelihood,
     ParametricLifetimeModel,
     document_args,
 )
 
 
-class LifetimeDistribution(ParametricLifetimeModel[()], ABC):
+class LifetimeDistribution(FittableParametricLifetimeModel[()], ABC):
     """
     Base class for distribution model.
     """
@@ -74,98 +78,32 @@ class LifetimeDistribution(ParametricLifetimeModel[()], ABC):
     def median(self) -> np.float64 | ArrayND[np.float64]:
         return self.ppf(0.5)  # no super here to return np.float64
 
-    @abstractmethod
-    def jac_hf(
-        self,
-        time: VT,
-    ) -> ArrayND[np.float64]:
-        """
-        The jacobian of the hazard function.
-
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.ndarray
-        """
-
-    @abstractmethod
-    def jac_chf(
-        self,
-        time: VT,
-    ) -> ArrayND[np.float64]:
-        """
-        The jacobian of the cumulative hazard function.
-
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.ndarray
-        """
-
-    @abstractmethod
-    def dhf(self, time: VT) -> ArrayND[np.float64]:
-        """
-        The derivate of the hazard function.
-
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.float64 or np.ndarray
-        """
-
+    @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_sf(self, time: VT) -> ArrayND[np.float64]:
-        """
-        The derivate of the survival function.
+        return super().jac_sf(time)
 
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.float64 or np.ndarray
-        """
-        jac_chf, sf = self.jac_chf(time), self.sf(time)
-        return -jac_chf * sf
-
+    @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_cdf(self, time: VT) -> ArrayND[np.float64]:
-        """
-        The derivate of the cumulative distribution function.
+        return super().jac_cdf(time)
 
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.float64 or np.ndarray
-        """
-        return -self.jac_sf(time)
-
+    @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_pdf(self, time: VT) -> ArrayND[np.float64]:
-        """
-        The derivate of the probability density function.
-
-        Parameters
-        ----------
-        time : float or np.ndarray
-
-        Returns
-        -------
-        out : np.float64 or np.ndarray
-        """
-        jac_hf, hf = self.jac_hf(time), self.hf(time)
-        jac_sf, sf = self.jac_sf(time), self.sf(time)
-        return jac_hf * sf + jac_sf * hf
+        return super().jac_pdf(time)
 
     @override
     @document_args(base_cls=ParametricLifetimeModel, args_docstring=[])
@@ -219,6 +157,31 @@ class LifetimeDistribution(ParametricLifetimeModel[()], ABC):
     def mrl(self, time: VT) -> np.float64 | ArrayND[np.float64]:
         return super().mrl(time)
 
+    @override
+    def init_likelihood(
+        self,
+        time: Array1D[np.float64] | Array[tuple[int, Literal[2]], np.float64],
+        args: Sequence[Array1D[np.float64]] | None = None,
+        event: Array1D[np.bool_] | None = None,
+        entry: Array1D[np.float64] | None = None,
+        **kwargs: Any,
+    ) -> LifetimeLikelihood:
+        assert args is None
+        lifetime_data = LifetimeData(time, event=event, entry=entry)
+        fresh_distrib = type(self)()
+        x0 = kwargs.get(
+            "x0", init_distrib_params_from_lifetimes(fresh_distrib, lifetime_data)
+        )
+        config = FitConfig(x0)
+        config.scipy_minimize_options["bounds"] = kwargs.get(
+            "bounds", get_distrib_params_bounds(fresh_distrib)
+        )
+        config.scipy_minimize_options["method"] = kwargs.get("method", "L-BFGS-B")
+        config.covariance_method = kwargs.get(
+            "covariance_method", "2point" if isinstance(fresh_distrib, Gamma) else "cs"
+        )
+        return LifetimeLikelihood(fresh_distrib, lifetime_data, config)
+
     def fit(
         self,
         time: Array1D[np.float64] | Array[tuple[int, Literal[2]], np.float64],
@@ -227,13 +190,40 @@ class LifetimeDistribution(ParametricLifetimeModel[()], ABC):
         **kwargs: Any,
     ) -> Self:
 
-        from relife.likelihoods import lifetime_likelihood
-
-        optimizer = lifetime_likelihood(self, time, event=event, entry=entry, **kwargs)
+        optimizer = self.init_likelihood(time, event=event, entry=entry, **kwargs)
         self.fitting_results = optimizer.optimize()
         self.set_params(self.fitting_results.optimal_params)
 
         return self
+
+
+def init_distrib_params_from_lifetimes(
+    model: LifetimeDistribution, data: LifetimeData
+) -> Array1D[np.float64]:
+    # flatten censored_time in case it is 2D
+    all_time_values = np.concatenate(
+        (data.complete_time.flatten(), data.censored_time.flatten())
+    )
+    nb_params = model.get_params().size
+    if isinstance(model, Gompertz):
+        param0 = np.empty(nb_params, dtype=np.float64)
+        rate = np.pi / (np.sqrt(6) * np.std(all_time_values))
+        shape = np.exp(-rate * np.mean(all_time_values))
+        param0[0] = shape
+        param0[1] = rate
+        return param0
+
+    param0 = np.ones(nb_params, dtype=np.float64)
+    param0[-1] = 1 / np.median(all_time_values)
+    return param0
+
+
+def get_distrib_params_bounds(model: LifetimeDistribution) -> Bounds:
+    nb_params = model.get_params().size
+    return Bounds(
+        np.full(nb_params, np.finfo(float).resolution),
+        np.full(nb_params, np.inf),
+    )
 
 
 @final
@@ -285,6 +275,11 @@ class Exponential(LifetimeDistribution):
         return cumulative_hazard_rate / self.get_params()[0]
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_hf(self, time: VT) -> ArrayND[np.float64]:
         if isinstance(time, np.ndarray):
             jac = np.expand_dims(np.ones_like(time, dtype=np.float64), axis=0)
@@ -293,6 +288,11 @@ class Exponential(LifetimeDistribution):
         return jac
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_chf(self, time: VT) -> ArrayND[np.float64]:
         if isinstance(time, np.ndarray):
             jac = np.expand_dims(time, axis=0).astype(np.float64)
@@ -301,6 +301,11 @@ class Exponential(LifetimeDistribution):
         return jac
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def dhf(self, time: VT) -> ArrayND[np.float64]:
         if isinstance(time, np.ndarray):
             return np.zeros_like(time, dtype=np.float64)
@@ -382,6 +387,11 @@ class Weibull(LifetimeDistribution):
         return np.asarray(cumulative_hazard_rate) ** (1 / shape) / rate
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_hf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.stack(
@@ -392,6 +402,11 @@ class Weibull(LifetimeDistribution):
         )
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_chf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.stack(
@@ -402,6 +417,11 @@ class Weibull(LifetimeDistribution):
         )
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def dhf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.asarray(
@@ -492,6 +512,11 @@ class Gompertz(LifetimeDistribution):
         return 1 / rate * np.log1p(cumulative_hazard_rate / shape)
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_hf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.stack(
@@ -502,6 +527,11 @@ class Gompertz(LifetimeDistribution):
         )
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_chf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.stack(
@@ -512,6 +542,11 @@ class Gompertz(LifetimeDistribution):
         )
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def dhf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return shape * rate**2 * np.exp(rate * time)
@@ -616,6 +651,11 @@ class Gamma(LifetimeDistribution):
         )
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_hf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         x = rate * time
@@ -631,6 +671,11 @@ class Gamma(LifetimeDistribution):
         return np.stack(jac)
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_chf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         x = rate * time
@@ -641,6 +686,11 @@ class Gamma(LifetimeDistribution):
         return np.stack(jac)
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def dhf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         return np.asarray(
@@ -721,6 +771,11 @@ class LogLogistic(LifetimeDistribution):
         return ((np.exp(cumulative_hazard_rate) - 1) ** (1 / shape)) / rate
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_hf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         x = rate * time
@@ -732,6 +787,11 @@ class LogLogistic(LifetimeDistribution):
         return np.stack(jac)
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def jac_chf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         x = rate * time
@@ -742,6 +802,11 @@ class LogLogistic(LifetimeDistribution):
         return np.stack(jac)
 
     @override
+    @document_args(
+        base_cls=FittableParametricLifetimeModel,
+        args_docstring=[],
+        returns=[docscrape.Parameter("out", "np.float64", [""])],
+    )
     def dhf(self, time: VT) -> ArrayND[np.float64]:
         shape, rate = self.get_params()
         x = rate * np.asarray(time)
