--- src/relife/likelihoods/__init__.py
@@ -1,4 +1,0 @@
-from ._cox_likelihood import cox_likelihood
-from ._lifetime_likelihood import lifetime_likelihood
-
-__all__: list[str] = ["lifetime_likelihood", "cox_likelihood"]
